Tweaks for editor log in (BL-12025)
